+++ sym/src/chacha20_avx512.c
@@ -32,7 +32,11 @@
 static int cc20_crypt_avx512(struct lc_sym_state *ctx, const uint8_t *in,
 			     uint8_t *out, size_t len)
 {
-	int ret;
+	int ret = cc20_check_overflow(ctx, len);
+
+	if (ret)
+		return ret;
+
 
 	/*
 	 * cc20_crypt_bytes_avx512 can handle the partial blocks, but we