+++ sym/src/chacha20_riscv64_v_zbb.c
@@ -32,22 +32,27 @@ static int cc20_crypt_riscv64_v_zbb(struct lc_sym_state *ctx,
 				    const uint8_t *in, uint8_t *out,
 				    size_t len)
 {
+	int ret;
+
 	LC_VECTOR_ENABLE;
-	cc20_crypt_asm(ctx, in, out, len, ChaCha20_riscv64_v_zbb);
+	ret = cc20_crypt_asm(ctx, in, out, len, ChaCha20_riscv64_v_zbb);
 	LC_VECTOR_DISABLE;
 
-	return 0;
+	return ret;
 }
 
 static int cc20_crypt_iv_riscv64_v_zbb(const struct lc_sym_state *ctx,
 				       const uint8_t *in, uint8_t *out,
 				       size_t len, uint8_t *iv, size_t ivlen)
 {
+	int ret;
+
 	LC_NEON_ENABLE;
-	cc20_crypt_iv_asm(ctx, in, out, len, iv, ivlen, ChaCha20_riscv64_v_zbb);
+	ret = cc20_crypt_iv_asm(ctx, in, out, len, iv, ivlen,
+				ChaCha20_riscv64_v_zbb);
 	LC_NEON_DISABLE;
 
-	return 0;
+	return ret;
 }
 
 static const struct lc_sym _lc_chacha20_riscv64_v_zbb = {